pyproject.toml
[tool.poetry]
name = "toxic-comments-mlops"
version = "0.1.0"
description = "MLOps project: multi-label toxic comment classification"
authors = ["Nikita Baglaev"]
readme = "README.md"
packages = [{ include = "toxic_comments" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Core
numpy = "^1.26.4"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
joblib = "^1.4.2"

# Deep learning
torch = "^2.3.1"
pytorch-lightning = "^2.3.3"
torchmetrics = "^1.4.0"
transformers = "^4.43.3"

# MLOps
hydra-core = "^1.3.2"
omgaconf = "^2.3.0"
mlflow = "^2.15.1"
dvc = "^3.55.2"
dvc-gdrive = "^3.0.1"

# Utils
tqdm = "^4.66.4"
matplotlib = "^3.9.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
